backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
sqlalchemy-cockroachdb==2.0.3
alembic==1.16.5
pydantic==2.11.9
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
httpx==0.28.1
# Use the modern psycopg package (successor to psycopg2)
psycopg[binary]==3.2.10
fast-agent-mcp==0.2.58
fastmcp==2.12.0
